Add changes to dom,logic,error handling
