+++ src/dom.js
@@ -1,31 +1,17 @@
 // PUSH TO DOM
-import Rainy from './assets/rain.jpg';
-import Fog from './assets/fog.jpg';
-import Cloud from './assets/cloud.jpg';
-import Sunny3 from './assets/suny3.jpg';
+import selectBakcground from './selectBackground';
+import { weatherTemp } from './logic';
 
 const pushToDom = (parsedWeather) => {
   // SELECTING BACKGROUND
-  const body = document.getElementById('body');
-  if (parsedWeather.weatherBackground === 'Clear') {
-    body.style.backgroundImage = `url(${Sunny3})`;
-  } else if (parsedWeather.weatherBackground === 'Rain' || parsedWeather.weatherBackground === 'Drizzle' || parsedWeather.weatherBackground === 'Thunderstorm') {
-    body.style.backgroundImage = `url(${Rainy})`;
-  } else if (parsedWeather.weatherBackground === 'Snow') {
-    body.style.backgroundImage = 'url("")';
-  } else if (parsedWeather.weatherBackground === 'Clouds') {
-    body.style.backgroundImage = `url(${Cloud})`;
-  } else {
-    body.style.backgroundImage = `url(${Fog})`;
-  }
-
+  selectBakcground(parsedWeather);
   // TEMP WEATHER BUTTON
   const tempButton = document.querySelector('.temp-button');
   tempButton.classList.remove('d-none');
   tempButton.classList.add('btn', 'btn-secondary', 'cel', 'my-3');
   tempButton.innerText = 'FAHRENHEIT';
 
-  // SELECTING AND CLEAN DOM
+  // SELECTING AND CLEAN selectBackgroundDOM
   const resultsDiv = document.querySelector('.content');
   const nameDiv = document.querySelector('.name');
   nameDiv.innerHTML = '';
@@ -73,33 +59,15 @@ const pushToDom = (parsedWeather) => {
 
   // TEMPERATURE BUTTON EVENT LISTENER
 
-  tempButton.addEventListener('click', (e) => {
-    if (e.target.classList.contains('cel')) {
-      tempButton.classList.remove('cel');
-      tempDiv.innerHTML = '';
-      const innerTempName = document.createElement('p');
-      innerTempName.textContent = 'Temperature';
-      const innerTempElement = document.createElement('p');
-      innerTempElement.textContent = parsedWeather.fCondition;
-      innerTempElement.innerHTML += '&#8457';
-      tempDiv.append(innerTempName, innerTempElement);
-
-      realTempDiv.innerHTML = '';
-      const innerRealTempName = document.createElement('p');
-      innerRealTempName.textContent = 'Real Thermal Feeling';
-      const innerRealTempElement = document.createElement('p');
-      innerRealTempElement.textContent = parsedWeather.fCondition;
-      innerRealTempElement.innerHTML += '&#8457';
-      realTempDiv.append(innerRealTempName, innerRealTempElement);
-      tempButton.innerText = 'CELSIUS';
-    } else {
-      tempButton.classList.add('cel');
-      tempDiv.innerHTML = '';
-      realTempDiv.innerHTML = '';
-      tempDiv.append(tempName, tempElement);
-      realTempDiv.append(realTempName, realTempElement);
-      tempButton.innerText = 'FAHRENHEIT';
-    }
+  tempButton.addEventListener('click', () => {
+    weatherTemp(tempButton,
+      tempDiv,
+      realTempDiv,
+      tempName,
+      tempElement,
+      realTempElement,
+      realTempName,
+      parsedWeather);
   });
 };
